Changed package.json to remove blessed and blessed-contrib as package dependancies and rather move it to temperature-control/package.json as an example dependancy to reduce bloat by ~14mb
