+++ examples/commonLib/cmdline.js
