+++ examples/device-example/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "device-example",
+  "version": "1.0.0",
+  "description": "",
+  "main": "device-example.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
