+++ examples/temperature-control/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "temperature-control",
+  "version": "1.0.0",
+  "description": "",
+  "main": "temperature-control.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "blessed": "^0.1.81",
+    "blessed-contrib": "^2.4.1"
+  }
+}
