+++ examples/thing-passthrough-example/thing-passthrough-example.js
@@ -18,8 +18,8 @@
 //npm deps
 
 //app deps
-const thingShadow = require('..').thingShadow;
-const cmdLineProcess   = require('./lib/cmdline');
+const thingShadow = require('../..').thingShadow;
+const cmdLineProcess   = require('../commonLib/cmdline');
 
 //begin module
